MRRTF-161: move MCH EntropyDecoder out of MCHWorkflow lib
There's no need for the ctf reader to depend on the (too wide for the
moment) full MCHWorkflow library.
