+++ Detectors/MUON/MCH/CTF/include/MCHCTF/EntropyDecoderSpec.h
